Added React.context api
Built a context provider component for the data i will be using inside of my routes, this will allow me to have the data all stored in one places. Cleaner code and also it will allow for greater functionality
This will also allow me to free up my app.js to be more of a hub for my routes and not needing to control so much of the props.